requirements.txt
pytest==8.3.5
requests==2.32.3
allure-pytest==2.13.5
pytest-asyncio==0.25.3

